Fixed frame transparency and moved FrameBlock class
