+++ src/main/java/com/simibubi/create/content/contraptions/components/structureMovement/chassis/FrameBlock.java
@@ -1,11 +1,11 @@
-package com.simibubi.create.content.contraptions.components.structureMovement.frame;
+package com.simibubi.create.content.contraptions.components.structureMovement.chassis;
 
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 
 public class FrameBlock extends Block {
 	public FrameBlock(Properties p_i48440_1_) {
-		super(p_i48440_1_);
+		super(p_i48440_1_.nonOpaque());
 	}
 
 	@Override
